% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nominatimlite-package.R
\docType{package}
\encoding{UTF-8}
\name{nominatimlite-package}
\alias{nominatimlite}
\alias{nominatimlite-package}
\title{nominatimlite: Interface to the 'Nominatim' API}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides a lightweight interface to the 'Nominatim' API \url{https://nominatim.org/release-docs/latest/}. It supports free-form and structured address searches, searches for addresses from coordinates, amenity lookup and address lookup by 'OpenStreetMap' object identifier. It returns results as 'tibble' data frames or 'sf' objects.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://dieghernan.github.io/nominatimlite/}
  \item \url{https://github.com/dieghernan/nominatimlite}
  \item Report bugs at \url{https://github.com/dieghernan/nominatimlite/issues}
}

}
\author{
\strong{Maintainer}: Diego Hernangómez \email{diego.hernangomezherrero@gmail.com} (\href{https://orcid.org/0000-0001-8457-4658}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Diego Hernangómez \email{diego.hernangomezherrero@gmail.com} (\href{https://orcid.org/0000-0001-8457-4658}{ORCID}) [copyright holder]
}

Other contributors:
\itemize{
  \item Jindra Lacko (\href{https://orcid.org/0000-0002-0375-5156}{ORCID}) [contributor, reviewer]
  \item Alex White [contributor]
  \item OpenStreetMap contributors (For the osm_amenities data) [copyright holder]
}

}
\keyword{internal}
