% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_address_lookup.R
\encoding{UTF-8}
\name{geo_address_lookup}
\alias{geo_address_lookup}
\title{Address lookup API}
\usage{
geo_address_lookup(
  osm_ids,
  type = c("N", "W", "R"),
  lat = "lat",
  long = "lon",
  full_results = FALSE,
  return_addresses = TRUE,
  verbose = FALSE,
  nominatim_server = "https://nominatim.openstreetmap.org/",
  custom_query = list()
)
}
\arguments{
\item{osm_ids}{A numeric vector of OSM identifiers, for example
\code{c(12345, 67890)}.}

\item{type}{Character vector of the OSM object type associated with each
\code{osm_ids} value. Possible values are node (\code{"N"}), way (\code{"W"}) or
relation (\code{"R"}). If a single value is provided, it will be recycled.}

\item{lat}{Name of the latitude column in the output. Defaults to \code{"lat"}.}

\item{long}{Name of the longitude column in the output. Defaults to \code{"lon"}.}

\item{full_results}{If \code{TRUE}, return all available fields from the Nominatim
API. If \code{FALSE}, return only query metadata, location data and requested
address columns.}

\item{return_addresses}{If \code{TRUE}, include single-line addresses in the
results.}

\item{verbose}{If \code{TRUE}, display detailed messages in the console.}

\item{nominatim_server}{Base URL of the Nominatim server. Defaults to
\code{"https://nominatim.openstreetmap.org/"}.}

\item{custom_query}{A named list of additional API parameters, for example
\code{list(countrycodes = "US")}. See \strong{Details}.}
}
\value{
A \code{\link[dplyr:tibble]{tibble}} with the results that match the query.
}
\description{
Looks up addresses and other details for one or more OpenStreetMap (OSM)
objects, such as nodes, ways or relations. Results are returned as a
\link[dplyr:tibble]{tibble}. Use \code{\link[=geo_address_lookup_sf]{geo_address_lookup_sf()}} to return an
\code{\link[sf:st_sf]{sf}} object instead.
}
\details{
See \url{https://nominatim.org/release-docs/latest/api/Lookup/} for additional
parameters to be passed to \code{custom_query}.
}
\examples{
\dontshow{if (nominatim_check_access()) withAutoprint(\{ # examplesIf}
\donttest{
ids <- geo_address_lookup(osm_ids = c(46240148, 34633854), type = "W")

ids

several <- geo_address_lookup(c(146656, 240109189), type = c("R", "N"))
several
}
\dontshow{\}) # examplesIf}
}
\seealso{
Address lookup functions:
\code{\link[=geo_address_lookup_sf]{geo_address_lookup_sf()}}

Address search functions:
\code{\link[=geo_address_lookup_sf]{geo_address_lookup_sf()}},
\code{\link[=geo_amenity]{geo_amenity()}},
\code{\link[=geo_amenity_sf]{geo_amenity_sf()}},
\code{\link[=geo_lite]{geo_lite()}},
\code{\link[=geo_lite_sf]{geo_lite_sf()}},
\code{\link[=geo_lite_struct]{geo_lite_struct()}},
\code{\link[=geo_lite_struct_sf]{geo_lite_struct_sf()}}
}
\concept{geocoding}
\concept{lookup}
