% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{geo_amenity}
\alias{geo_amenity}
\alias{geo_amenity_sf}
\title{Geocode amenities}
\usage{
geo_amenity(bbox = NULL, ...)

geo_amenity_sf(bbox = NULL, ...)
}
\arguments{
\item{bbox, ...}{Deprecated}
}
\value{
An error.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#defunct}{\figure{lifecycle-defunct.svg}{options: alt='[Defunct]'}}}{\strong{[Defunct]}}

This operation is not supported any more. Use
\code{\link[arcgeocoder:arc_geo_categories]{arcgeocoder::arc_geo_categories()}} instead.
}
\examples{
\donttest{
# Madrid, Spain

library(arcgeocoder)
library(ggplot2)

bbox <- c(-3.888954, 40.311977, -3.517916, 40.643729)

# Food
rest_pub <- arc_geo_categories(
  bbox = bbox, category = "Bakery,Bar or Pub",
  full_results = TRUE,
  limit = 50
)

rest_pub
}
}
\keyword{internal}
