% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_lite_sf.R
\encoding{UTF-8}
\name{geo_lite_sf}
\alias{geo_lite_sf}
\title{Address search API with \CRANpkg{sf} output (free-form query)}
\usage{
geo_lite_sf(
  address,
  limit = 1,
  return_addresses = TRUE,
  full_results = FALSE,
  verbose = FALSE,
  progressbar = TRUE,
  nominatim_server = "https://nominatim.openstreetmap.org/",
  custom_query = list(),
  points_only = TRUE
)
}
\arguments{
\item{address}{\code{character} with a single-line address, for example
\code{"1600 Pennsylvania Ave NW, Washington"}, or a vector of addresses
(\code{c("Madrid", "Barcelona")}).}

\item{limit}{Maximum number of results to return per input address. Note
that each query returns a maximum of 50 results.}

\item{return_addresses}{Return input addresses with results if \code{TRUE}.}

\item{full_results}{Return all available data from the Nominatim API.
If \code{FALSE} (default), only address columns are returned. See also
\code{return_addresses}.}

\item{verbose}{If \code{TRUE}, detailed logs are output to the console.}

\item{progressbar}{Logical. If \code{TRUE} displays a progress bar to indicate
the progress of the function.}

\item{nominatim_server}{URL of the Nominatim server to use. Defaults to
\code{"https://nominatim.openstreetmap.org/"}.}

\item{custom_query}{Named list with API-specific parameters, for example
\code{list(countrycodes = "US")}. See \strong{Details}.}

\item{points_only}{Logical \code{TRUE/FALSE}. Whether to return only point
geometries (\code{TRUE}, which is the default) or potentially other shapes as
returned by the Nominatim API (\code{FALSE}). See \strong{About geometry types}.}
}
\value{
An \code{\link[sf:st_sf]{sf}} object with the results that match the query.
}
\description{
Geocodes addresses and returns the corresponding \code{\link[sf:st_sf]{sf}} object.
The query output is returned as an \CRANpkg{sf} object. See \code{\link[=geo_lite]{geo_lite()}} for
retrieving the data in \code{\link[tibble:tibble]{tibble}} format.

Corresponds to the \strong{free-form query} search described in the
\href{https://nominatim.org/release-docs/latest/api/Search/}{API endpoint}.
}
\details{
See \url{https://nominatim.org/release-docs/latest/api/Search/} for additional
parameters to be passed to \code{custom_query}.
}
\section{About geometry types}{


The parameter \code{points_only} specifies whether the function results will be
points (all Nominatim results are guaranteed to have at least point
geometry) or other geometry types.

Note that when \code{points_only = FALSE}, the type of geometry returned depends
on the object being geocoded. Administrative areas, major buildings and the
like will be returned as polygons, rivers, roads and similar features will
be returned as lines, and amenities may still be returned as points.

This function is vectorized, allowing multiple addresses to be geocoded.
With \code{points_only = FALSE}, multiple geometry types may be returned.
}

\examples{
\dontshow{if (nominatim_check_access()) withAutoprint(\{ # examplesIf}
\donttest{
# Map: points
library(ggplot2)

string <- "Statue of Liberty, NY, USA"
sol <- geo_lite_sf(string)

if (!all(sf::st_is_empty(sol))) {
  ggplot(sol) +
    geom_sf()
}

sol_poly <- geo_lite_sf(string, points_only = FALSE)

if (!all(sf::st_is_empty(sol_poly))) {
  ggplot(sol_poly) +
    geom_sf() +
    geom_sf(data = sol, color = "red")
}
# Several results

madrid <- geo_lite_sf("Comunidad de Madrid, Spain",
  limit = 2,
  points_only = FALSE, full_results = TRUE
)

if (!all(sf::st_is_empty(madrid))) {
  ggplot(madrid) +
    geom_sf(fill = NA)
}
}
\dontshow{\}) # examplesIf}
}
\seealso{
Geocoding functions:
\code{\link[=geo_address_lookup]{geo_address_lookup()}},
\code{\link[=geo_address_lookup_sf]{geo_address_lookup_sf()}},
\code{\link[=geo_amenity]{geo_amenity()}},
\code{\link[=geo_amenity_sf]{geo_amenity_sf()}},
\code{\link[=geo_lite]{geo_lite()}},
\code{\link[=geo_lite_struct]{geo_lite_struct()}},
\code{\link[=geo_lite_struct_sf]{geo_lite_struct_sf()}}

Spatial output functions:
\code{\link[=bbox_to_poly]{bbox_to_poly()}},
\code{\link[=geo_address_lookup_sf]{geo_address_lookup_sf()}},
\code{\link[=geo_amenity_sf]{geo_amenity_sf()}},
\code{\link[=geo_lite_struct_sf]{geo_lite_struct_sf()}},
\code{\link[=reverse_geo_lite_sf]{reverse_geo_lite_sf()}}
}
\concept{geocoding}
\concept{spatial}
