% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_lite_struct_sf.R
\encoding{UTF-8}
\name{geo_lite_struct_sf}
\alias{geo_lite_struct_sf}
\title{Address search API with \CRANpkg{sf} output (structured query)}
\usage{
geo_lite_struct_sf(
  amenity = NULL,
  street = NULL,
  city = NULL,
  county = NULL,
  state = NULL,
  country = NULL,
  postalcode = NULL,
  limit = 1,
  full_results = FALSE,
  return_addresses = TRUE,
  verbose = FALSE,
  nominatim_server = "https://nominatim.openstreetmap.org/",
  custom_query = list(),
  points_only = TRUE
)
}
\arguments{
\item{amenity}{Name and/or type of POI. See also \code{\link[=geo_amenity]{geo_amenity()}}.}

\item{street}{House number and street name.}

\item{city}{City.}

\item{county}{County.}

\item{state}{State.}

\item{country}{Country.}

\item{postalcode}{Postal code.}

\item{limit}{Maximum number of results to return per input address. Note
that each query returns a maximum of 50 results.}

\item{full_results}{Return all available data from the Nominatim API.
If \code{FALSE} (default), only latitude, longitude and address columns are
returned. See also \code{return_addresses}.}

\item{return_addresses}{Return input addresses with results if \code{TRUE}.}

\item{verbose}{If \code{TRUE}, detailed logs are output to the console.}

\item{nominatim_server}{URL of the Nominatim server to use. Defaults to
\code{"https://nominatim.openstreetmap.org/"}.}

\item{custom_query}{Named list with API-specific parameters, for example
\code{list(countrycodes = "US")}. See \strong{Details}.}

\item{points_only}{Logical \code{TRUE/FALSE}. Whether to return only point
geometries (\code{TRUE}, which is the default) or potentially other shapes as
returned by the Nominatim API (\code{FALSE}). See \strong{About geometry types}.}
}
\value{
An \code{\link[sf:st_sf]{sf}} object with the results that match the query.
}
\description{
Geocodes addresses already split into components and returns the
corresponding \code{\link[sf:st_sf]{sf}} object. The query output is returned as an
\CRANpkg{sf} format. See \code{\link[=geo_lite_struct]{geo_lite_struct()}} for retrieving the data in
\code{\link[tibble:tibble]{tibble}} format.

Corresponds to the \strong{structured query} search described in the
\href{https://nominatim.org/release-docs/latest/api/Search/}{API endpoint}. To
perform a free-form search, use \code{\link[=geo_lite_sf]{geo_lite_sf()}}.
}
\details{
The structured form of the search query allows you to look up an address that
is already split into its components. Each parameter represents a field of
the address. All parameters are optional. You should only use the ones that
are relevant for the address you want to geocode.

See \url{https://nominatim.org/release-docs/latest/api/Search/} for additional
parameters to be passed to \code{custom_query}.
}
\section{About geometry types}{


The parameter \code{points_only} specifies whether the function results will be
points (all Nominatim results are guaranteed to have at least point
geometry) or other geometry types.

Note that when \code{points_only = FALSE}, the type of geometry returned depends
on the object being geocoded. Administrative areas, major buildings and the
like will be returned as polygons, rivers, roads and similar features will
be returned as lines, and amenities may still be returned as points.

This function is vectorized, allowing multiple addresses to be geocoded.
With \code{points_only = FALSE}, multiple geometry types may be returned.
}

\examples{
\dontshow{if (nominatim_check_access()) withAutoprint(\{ # examplesIf}
\donttest{
# Map

pl_mayor <- geo_lite_struct_sf(
  street = "Plaza Mayor",
  county = "Comunidad de Madrid",
  country = "Spain", limit = 50,
  full_results = TRUE, verbose = TRUE
)

# Outline
ccaa <- geo_lite_sf("Comunidad de Madrid, Spain", points_only = FALSE)

library(ggplot2)

if (any(!sf::st_is_empty(pl_mayor), !sf::st_is_empty(ccaa))) {
  ggplot(ccaa) +
    geom_sf() +
    geom_sf(data = pl_mayor, aes(shape = addresstype, color = addresstype))
}
}
\dontshow{\}) # examplesIf}
}
\seealso{
Geocoding:
\code{\link[=geo_address_lookup]{geo_address_lookup()}},
\code{\link[=geo_address_lookup_sf]{geo_address_lookup_sf()}},
\code{\link[=geo_amenity]{geo_amenity()}},
\code{\link[=geo_amenity_sf]{geo_amenity_sf()}},
\code{\link[=geo_lite]{geo_lite()}},
\code{\link[=geo_lite_sf]{geo_lite_sf()}},
\code{\link[=geo_lite_struct]{geo_lite_struct()}}

\code{\link[sf:sf]{sf}} outputs:
\code{\link[=bbox_to_poly]{bbox_to_poly()}},
\code{\link[=geo_address_lookup_sf]{geo_address_lookup_sf()}},
\code{\link[=geo_amenity_sf]{geo_amenity_sf()}},
\code{\link[=geo_lite_sf]{geo_lite_sf()}},
\code{\link[=reverse_geo_lite_sf]{reverse_geo_lite_sf()}}
}
\concept{geocoding}
\concept{spatial}
