% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nominatim_check_access.R
\encoding{UTF-8}
\name{nominatim_check_access}
\alias{nominatim_check_access}
\title{Check access to the Nominatim API}
\usage{
nominatim_check_access(
  nominatim_server = "https://nominatim.openstreetmap.org/"
)
}
\arguments{
\item{nominatim_server}{A string giving the base URL of the Nominatim
server. Defaults to
\code{"https://nominatim.openstreetmap.org/"}.}
}
\value{
A single logical value: \code{TRUE} if the API is available and \code{FALSE} otherwise.
}
\description{
Checks whether \R can access a Nominatim API server.
}
\examples{
\donttest{
nominatim_check_access()
}
}
\seealso{
\url{https://nominatim.org/release-docs/latest/api/Status/}.
}
\concept{API management}
\keyword{internal}
