% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nominatim_check_access.R
\name{nominatim_check_access}
\alias{nominatim_check_access}
\title{Check access to Nominatim API}
\usage{
nominatim_check_access()
}
\value{
a logical.
}
\description{
Check if \strong{R} has access to resources at
\url{https://nominatim.openstreetmap.org}.
}
\examples{
\donttest{
nominatim_check_access()
}
}
\seealso{
\url{https://nominatim.org/release-docs/latest/api/Status/}

API Management: 
\code{\link{api_call}()}
}
\concept{api_management}
\keyword{internal}
