% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nominatim_check_access.R
\name{nominatim_check_access}
\alias{nominatim_check_access}
\title{Check access to Nominatim API}
\usage{
nominatim_check_access()
}
\value{
a logical.
}
\description{
Check if R has access to resources at
\url{https://nominatim.openstreetmap.org}.
}
\examples{

nominatim_check_access()
}
\concept{helper}
\keyword{internal}
