% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocoderlite-package.R
\docType{package}
\name{nominatimlite-package}
\alias{nominatimlite}
\alias{nominatimlite-package}
\title{nominatimlite: Interface with 'Nominatim' API Service}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Lite interface for getting data from 'OSM' service
    'Nominatim'.  See <https://nominatim.org/release-docs/latest/>. This
    is a lite version of the package 'tidygeocoder'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://dieghernan.github.io/nominatimlite/}
  \item \url{https://github.com/dieghernan/nominatimlite}
  \item Report bugs at \url{https://github.com/dieghernan/nominatimlite/issues}
}

}
\author{
\strong{Maintainer}: Diego Hernangómez \email{diego.hernangomezherrero@gmail.com} (\href{https://orcid.org/0000-0001-8457-4658}{ORCID}) [copyright holder]

Other contributors:
\itemize{
  \item Jesse Cambon (\href{https://orcid.org/0000-0001-6854-1514}{ORCID}) [contributor]
  \item OpenStreetMap contributors (for the data.) [copyright holder]
}

}
\keyword{internal}
