% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nominatimlite-package.R
\docType{package}
\name{nominatimlite-package}
\alias{nominatimlite}
\alias{nominatimlite-package}
\title{nominatimlite: Interface with 'Nominatim' API Service}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Lite interface for getting data from 'OSM' service 'Nominatim' \url{https://nominatim.org/release-docs/latest/}. Extract coordinates from addresses, find places near a set of coordinates and return spatial objects on 'sf' format.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://dieghernan.github.io/nominatimlite/}
  \item \url{https://github.com/dieghernan/nominatimlite}
  \item Report bugs at \url{https://github.com/dieghernan/nominatimlite/issues}
}

}
\author{
\strong{Maintainer}: Diego Hernangómez \email{diego.hernangomezherrero@gmail.com} (\href{https://orcid.org/0000-0001-8457-4658}{ORCID}) [copyright holder]

Other contributors:
\itemize{
  \item Jindra Lacko (\href{https://orcid.org/0000-0002-0375-5156}{ORCID}) [contributor, reviewer]
  \item Alex White [contributor]
}

}
\keyword{internal}
