% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{osm_amenities}
\alias{osm_amenities}
\title{OpenStreetMap amenities}
\format{
A \link[dplyr:tibble]{tibble} with
140 rows and
three columns:
\describe{
\item{category}{Amenity category.}
\item{amenity}{Amenity value.}
\item{comment}{Brief description of the amenity type.}
}
}
\source{
\url{https://wiki.openstreetmap.org/wiki/Key:amenity}
}
\description{
A dataset of amenity values available on OpenStreetMap.
}
\note{
The data were extracted on \strong{July 11, 2026}. See
\code{inst/COPYRIGHTS} for copyright and license details.
}
\examples{
data("osm_amenities")

osm_amenities
}
\seealso{
Amenity lookup functions and data:
\code{\link[=geo_amenity]{geo_amenity()}},
\code{\link[=geo_amenity_sf]{geo_amenity_sf()}}
}
\concept{amenity}
\keyword{datasets}
